refactor chroma parameter validation fix chroma tests not running sequentially
